requirements.txt
selenium
pytest
openpyxl
allure-pytest